test: extract exploration tool white-box suite
