+++ scripts/rust-source-size-baseline.txt
@@ -20,5 +20,5 @@ crates/shell-core/src/control.rs 1183
 crates/shell-core/src/model.rs 1320
 crates/shell-core/src/reducer.rs 2723
 crates/shell-core/src/terminal_profile.rs 1852
-crates/tools/src/lib.rs 1459
+crates/tools/src/lib.rs 1014
 crates/tui/src/file_backed.rs 2753
